+++ numunsigned.go
@@ -42,13 +42,13 @@ func (b UnsignedBuilder[T]) WithDefault(v T) UnsignedBuilder[T] {
 // Required completes the build process and registers a required variable with
 // Ferrite's validation system.
 func (b UnsignedBuilder[T]) Required() Required[T] {
-	return registerRequired(b.spec(), b.resolve)
+	return spec.RegisterRequired(b.spec(), b.resolve)
 }
 
 // Optional completes the build process and registers an optional variable with
 // Ferrite's validation system.
 func (b UnsignedBuilder[T]) Optional() Optional[T] {
-	return registerOptional(b.spec(), b.resolve)
+	return spec.RegisterOptional(b.spec(), b.resolve)
 }
 
 func (b UnsignedBuilder[T]) spec() spec.Spec {