+++ string.go
@@ -45,13 +45,13 @@ func (b StringBuilder[T]) WithDefault(v T) StringBuilder[T] {
 // Required completes the build process and registers a required variable with
 // Ferrite's validation system.
 func (b StringBuilder[T]) Required() Required[T] {
-	return registerRequired(b.spec(), b.resolve)
+	return spec.RegisterRequired(b.spec(), b.resolve)
 }
 
 // Optional completes the build process and registers an optional variable with
 // Ferrite's validation system.
 func (b StringBuilder[T]) Optional() Optional[T] {
-	return registerOptional(b.spec(), b.resolve)
+	return spec.RegisterOptional(b.spec(), b.resolve)
 }
 
 func (b StringBuilder[T]) spec() spec.Spec {